+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 4.0 [not yet released]
+
+- added smoothness encoded value, used to determine bike speed
+
 ### 3.0 [17 May 2021]
 
 - removed the stable tag (was pointing to commit dd2c20c763e4c19b701e92386432b37713cd8dc5)
@@ -304,8 +308,7 @@
 
 ### 0.2.0 [23 Nov 2013]
 
-- change inconsistent default settings for contraction hierarchies in the API -
-  see https://lists.openstreetmap.org/pipermail/graphhopper/2013-December/000585.html
+- change inconsistent default settings for contraction hierarchies in the API - see https://lists.openstreetmap.org/pipermail/graphhopper/2013-December/000585.html
 - fixed issues with android:
   * graphhopper: use maps from 0.2 path; updated maps
   * mapsforge: use mapsforge-map dependency; merged #461; avoid duplicates otherwise mapsforge-core would be duplicate (